requirements.txt
ccxt
pandas
numpy
streamlit
python-dotenv
plotly
watchdog
git+https://github.com/twopirllc/pandas-ta.git@development
